Read RUN_APPLICATION from sdk
