+++ src_plugin_sdk/main.c
@@ -17,6 +17,7 @@
 
 #include "eth_internals.h"
 #include "eth_plugin_interface.h"
+#include "lib_standard_app/swap_lib_calls.h"  // RUN_APPLICATION
 
 // Functions implemented by the plugin
 void handle_init_contract(ethPluginInitContract_t *parameters);
