+++ tools/build_sdk.py
@@ -161,7 +161,6 @@ if __name__ == "__main__":
         "src/shared_context.h",
         "src/eth_plugin_internal.h",
         "src/nft.h",
-        "src/swap_lib_calls.h",
     ]
     nodes_to_extract = {
         "#define": ["MAX_TICKER_LEN",
@@ -170,7 +169,6 @@ if __name__ == "__main__":
                     "WEI_TO_ETHER",
                     "SELECTOR_SIZE",
                     "PARAMETER_LENGTH",
-                    "RUN_APPLICATION",
                     "COLLECTION_NAME_MAX_LEN"],
         "typedef enum": [],
         "typedef struct": ["tokenDefinition_t",
